[DT-338] Extract version code and language interceptors
